All template code is in the heder file
