+++ opencog/util/CMakeLists.txt
@@ -18,7 +18,6 @@ ADD_DEFINITIONS(
 ADD_LIBRARY(cogutil SHARED
 	ansi
 	algorithm
-	async_writer
 	backtrace-symbols
 	based_variant
 	cluster
@@ -101,7 +100,7 @@ INSTALL(FILES
 	misc.h
 	mt19937ar.h
 	numeric.h
-    	jaccard_index.h
+	jaccard_index.h
 	oc_assert.h
 	oc_omp.h
 	octime.h
